Added a method to calculate a molecular orbital
This calls Python API added to AvogadroLibs, and so will require an
update. It also uses new code in the CJSON writer to output a cube in
JSON format that is very simple to bring into 3Dmol (and I imagine most
other JavaScript based viewers). It will likely change in future to
annotate what the cube actually represents, it must currently be
inferred and can only carry one cube.