Fix molecule server error if attributes missing
Some of these attributes are not required for molecules. For instance,
if a molecule is created via xyz, a name will not necessarily be
given to it.

If there are any molecules in the database without a name, GET
'/molecules' would cause a server error.

This fixes the server error.