Add start of InChI validator to prevent OB crashes
Add a start of InChI validator to prevent Open Babel crashes.
In my testing, this fixes most of the server crashes. I have
not yet found a way to make the server crash with an invalid
InChI while this validator is present. But we can add more to
the validator in the future if needed.

Fixes: #105

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>