Do not create a notebook if there is no assetstore
As of girder 3, in mongochemdeploy, when ansible creates the first user
`mongochemdev`, it also tries to create a notebook due to the notebook
plugin, and an exception gets raised because there is no assestore yet.

Before girder 3, this wouldn't happen because the notebook plugin would
be explicitly enabled after the `mongochemdev` user was created. But in
girder 3, the plugin is enabled automatically at the start if the
plugin was pip installed.

This commit prevents the exception from occurring in ansible.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>