Save a creatorId in the molecule
This will be the id of the user who created the molecule.

This is done so that one can search for molecules created
by a specific user.

Note that we needed to convert the id to a string, or else
we will not be able to find it by query in self.find()
(unless we convert the creatorId string in the query
to an ObjectId, I think).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>