Add molecule field to indicate 3d coords pending
Added a `generating_3d_coords` field to molecules that indicates
that 3d coordinates are currently being generated. This is not
currently being used, but it will be used in the future so that
3d coordinates will not be generated twice.

The field is removed after 3d coordinates are generated. Absense
of the field is indicative that 3d coordinates are not being
generated.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>