Renamed "Simultaneous Calcs" to "Multiple Calcs"
All of the calculations are ran in serial, so the name
"Simultaneous Calcs" might be misleading. Use the name
"Multiple Calcs" instead.

Also modified a note that an exception will be raised when
running a 3D calculation on a 2D molecule, because an exception
is no longer raised, but just a message printed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>