Generate SVG image asynchronously
Generating an SVG image can take a long time for big molecules
(sometimes, more than 1 minute).

Generating it asynchronously should, in theory, be a nice improvement
to server performance. However, the server is still unresponsive while
it is generating the SVG image asynchronously.

Perhaps Open Babel is not releasing the python GIL when it runs the SVG
generation code.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>