Asynchronously generate SVG with Open Babel flask
Added requests-futures as a dependency, and used it to make
asynchronous requests to generate SVG images. It seems to be
working fairly well. The Open Babel URL is now a setting in
molecules.constants.

Also renamed "generate_3d_coords_async.py" to a more general name,
"async_jobs.py".

Still need to start using a production server for the Open Babel
flask container.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>