When generating inchi, do not convert to xyz first
Converting to xyz has potentially been causing some issues with the
inchi generation. Perhaps it is because the molecule loses the
connectivity and needs to perceive it again - and all atoms get
placed at (0, 0, 0), so bond perception may not be working.

For instance, if we create an inchi from ethane, CC, but we convert
to xyz first, it places both carbon atoms at (0, 0, 0), and only
four hydrogen atoms are added around the two carbon atoms that are
sitting on top of each other. It produces this inchi and inchikey:
```
InChI=1S/2CH4/h2*1H4
CREMABGTGYGIQB-UHFFFAOYSA-N
```

Instead of what appears to be the correct inchi and inchikey:
```
InChI=1S/C2H6/c1-2/h1-2H3
OTMSDBZUPAUEDD-UHFFFAOYSA-N
```

Skipping the xyz conversion seems to fix the issue, and we get the
correct inchi.

However, we should perhaps test this a little more and make sure it
doesn't break anything.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>