Check bonds for 3D inputs and auto generate
This checks the bonding for 3D inputs to see if it either doesn't
have bonds or only has single bonds. If either of these are true,
it will automatically attempt to detect the bonds. This fixes cases
where molecules like caffeine, for instance, are reported as having
all single bonds. It is very important for us to get the bonding
correct, or else the InChI and InChI key will not be correct.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>