Move avogadro service to 5001
This change is really to support CircleCI but its probably not
a bad idea to separate the service ports. Basically CircleCI maps
all container ports to localhost on the primary docker container.
This means that avogadro and openbabel fight for 5000. This is not
a problem with docker-compose as it sets up separate interfaces for
each container.