Hash input params on server for calc search
This puts the input parameter hashing on the server instead of
on the client side for searching for a calculation. It prevents
clients from having to ensure they hash the input parameters in
the same way that the server does it.

It is intended to send this via url quoted data. However, it
seems to work fine without url quoting also... (probably as long
as there are no percent signs in the string).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>