Allow user to pass in name or wikipedia link
If a 'name' or 'wikipedia' field are passed in via the body of a molecule those
fields will be set. If 'name' or 'wikipedia' are included in the input parameters
of a calculation and the molecule does not yet exist those fields will be set
upon molecule creation. If the molecule already exists the fields are ignored.

Signed-off-by: Brianna Major <brianna.major@kitware.com>