Add forcefield and steps options for gen3d
This adds options to choose the forcefield (default: 'mmff94') and
the number of optimization steps (default: 100) when generating 3D
coordinates for a molecule.

In the POST /molecules REST API, the forcefield can be specified with
"gen3dForcefield", and the number of steps can be specified with
"gen3dSteps". In the POST /molecules/{id}/3d REST API, the forcefield
can be specified with the "forcefield" query, and the number of steps
can be specified with the "steps" query.

The previous number of optimization steps was 50. This commit bumps
the default number of optimization steps up to 100, so the molecules
will probably be slightly different than before.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>