Convert steps to int for generating 3D coordinates
For the end point POST /molecules/{id}/3d, the steps is a query, and
is thus treated as a string. Convert it to an int before calling
schedule_3d_coords_gen() with it.

This was not causing a problem with the web API since it the web API
automatically converts it to an int. However, it is a problem for
calling this end point in other contexts, including openchemistrypy.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>