Fix GET /calculations with a geometry ID
The geometry ID is not being stored as an ObjectId currently. Thus,
if we try to query it as an ObjectId, we find no results. Fix this
by removing the ObjectId conversion.

Another alternative to fix this is to convert the geometry ID into
an ObjectId when it gets saved on the calculation.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>