Add a single endpoint to launch taskflows
Rather than having ~7 restful API calls to launch a single taskflow,
add a single restful API endpoint that takes care of most of the logic.

This single API call optionally takes a clusterId. It will use that
cluster if provided. If not provided one, it will use the first cluster
it can find.

The body of the endpoint should contain two keys:

1. taskFlowBody
2. taskBody

That contain the bodies of the taskflow, and the task, respectively.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>